// fe/src/lib/api/generated/model/familyMemberKindValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type FamilyMemberKindValue = typeof FamilyMemberKindValue[keyof typeof FamilyMemberKindValue];


export const FamilyMemberKindValue = {
  FATHER: 'FATHER',
  MOTHER: 'MOTHER',
  SPOUSE: 'SPOUSE',
} as const;