// fe/src/lib/api/generated/model/immunizationValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ImmunizationValue = typeof ImmunizationValue[keyof typeof ImmunizationValue];


export const ImmunizationValue = {
  CHICKEN_POX: 'CHICKEN_POX',
  BOOSTER: 'BOOSTER',
  MEASLES_MMR: 'MEASLES_MMR',
  HEPATITIS_B: 'HEPATITIS_B',
  MUMPS: 'MUMPS',
  INFLUENZA: 'INFLUENZA',
  SMALL_POX: 'SMALL_POX',
  OTHER: 'OTHER',
} as const;